pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = ""
authors = [
    {name = "gihunKIM",email = "kkh1902@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyarrow (==21.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "google-cloud-storage (>=3.8.0,<4.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "notebook (>=7.5.2,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyter (>=1.1.1,<2.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]
